refactor(navigation): move SearchNavigation to feature-search to honor feature-owned deep links
- Move openSearch/decodeSearchFilter from navigation-api to feature-search
- Update imports in users and search to reference new location
- Keeps navigation-api generic; search owns its deep links as per standards.
